Move most of authentication from admin to common
